requirements.txt
geopandas>=0.14.0
shapely>=2.0.0
pyyaml>=6.0
PyPDF2>=3.0.0
pytest>=7.4.0

